requirements.txt
amqp==5.1.1
asgiref==3.5.2
async-timeout==4.0.2
attrs==21.4.0
billiard==3.6.4.0
celery==5.2.7
certifi==2022.5.18.1
charset-normalizer==2.0.12
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
coverage==6.4.1
Deprecated==1.2.13
Django==4.0.5
django-celery-beat==2.3.0
django-timezone-field==5.0
furl==2.1.3
idna==3.3
jsonschema==4.6.0
kombu==5.2.4
multidimensional-urlencode==0.0.4
nose==1.3.7
orderedmultidict==1.0.1
packaging==21.3
pook==1.0.2
prompt-toolkit==3.0.29
pyparsing==3.0.9
pyrsistent==0.18.1
python-crontab==2.6.0
python-dateutil==2.8.2
pytz==2022.1
redis==4.3.3
requests==2.27.1
retailcrm==5.1.2
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
urllib3==1.26.9
vine==5.0.0
wcwidth==0.2.5
wrapt==1.14.1
xmltodict==0.13.0
